requirements.txt
opencv-contrib-python
mediapipe
numpy
pillow
streamlit
streamlit-webrtc
